Include original PRO term name as EXACT synonym
We replace the PRO term name in the mini PRO OBO file with
the EXACT PomBase_display_name synonym.  Instead of discarding
the original term name we now save it as an EXACT synonym with
"PomBase_original_PRO_term_name: " as a prefix.

This is a hack to allow us to include the original term name in
column 3 of the GPI file.  :-(

Refs pombase/curation#2851
Refs pombase/pombase-chado#1173